(refactor) moved some components to single files
