+++ src/components/account/wallet_weth_modal.tsx
@@ -6,6 +6,8 @@ import styled from 'styled-components';
 import { themeColors } from '../../util/theme';
 import { tokenAmountInUnits, unitsInTokenAmount } from '../../util/tokens';
 import { Button as ButtonBase } from '../common/button';
+import { CloseModalButton } from '../common/close_modal_button';
+import { InfoIcon } from '../common/info_icon';
 
 interface Props extends React.ComponentProps<typeof Modal> {
     wethBalance: BigNumber;
@@ -57,30 +59,8 @@ const CloseButtonContainer = styled.div`
     display: flex;
     height: 20px;
     justify-content: flex-end;
-`;
-
-const CloseButton = styled.span`
-    align-items: center;
-    cursor: pointer;
-    display: flex;
-    height: 100%;
-    justify-content: center;
     margin-right: -10px;
-    width: 20px;
 `;
-
-const CloseButtonSVG = () => {
-    return (
-        <svg width="11" height="11" viewBox="0 0 11 11" fill="none" xmlns="http://www.w3.org/2000/svg">
-            <path
-                fillRule="evenodd"
-                clipRule="evenodd"
-                d="M10.4501 10.449C10.7538 10.1453 10.7538 9.65282 10.4501 9.34909L6.60062 5.49996L10.45 1.65093C10.7537 1.3472 10.7537 0.854765 10.45 0.551038C10.1462 0.247311 9.65374 0.24731 9.34999 0.551038L5.50063 4.40006L1.65018 0.549939C1.34643 0.246212 0.853943 0.246212 0.55019 0.549939C0.246437 0.853667 0.246436 1.34611 0.55019 1.64983L4.40064 5.49996L0.550081 9.35019C0.246327 9.65392 0.246327 10.1464 0.550081 10.4501C0.853834 10.7538 1.34632 10.7538 1.65007 10.4501L5.50063 6.59985L9.3501 10.449C9.65385 10.7527 10.1463 10.7527 10.4501 10.449Z"
-                fill="#C4C4C4"
-            />
-        </svg>
-    );
-};
 
 const Title = styled.h1`
     color: #000;
@@ -107,6 +87,14 @@ const EthBox = styled.div`
     justify-content: center;
     min-height: 105px;
     padding: 10px;
+    position: relative;
+`;
+
+const InfoIconStyled = styled.div`
+    cursor: pointer;
+    position: absolute;
+    right: 8px;
+    top: 8px;
 `;
 
 interface EthBoxProps {
@@ -140,9 +128,9 @@ const SetMinEthWrapper = styled.div`
 `;
 
 const SetMinEthButton = styled.a`
+    border-bottom: 1px dotted black;
     color: #666;
     text-decoration: none;
-    border-bottom: 1px dotted black;
 `;
 
 const minEth = unitsInTokenAmount('0.5', 18);
@@ -170,9 +158,7 @@ class WethModal extends React.Component<Props, State> {
         return (
             <Modal {...restProps}>
                 <CloseButtonContainer>
-                    <CloseButton onClick={this.closeModal}>
-                        <CloseButtonSVG />
-                    </CloseButton>
+                    <CloseModalButton onClick={this.closeModal} />
                 </CloseButtonContainer>
                 <Title>Available Balance</Title>
                 <EthBoxes>
@@ -183,6 +169,9 @@ class WethModal extends React.Component<Props, State> {
                     <EthBox>
                         <EthBoxValue isZero={selectedWethStr === minSlidervalue}>{selectedWethStr}</EthBoxValue>
                         <EthBoxUnit>wETH</EthBoxUnit>
+                        <InfoIconStyled>
+                            <InfoIcon />
+                        </InfoIconStyled>
                     </EthBox>
                 </EthBoxes>
                 <Slider
